[ODESolver] Rename files to match class name
