+++ Sofa/Component/ODESolver/Forward/CMakeLists.txt
@@ -7,6 +7,7 @@ set(HEADER_FILES
     ${SOFACOMPONENTODESOLVERFORWARD_SOURCE_DIR}/config.h.in
     ${SOFACOMPONENTODESOLVERFORWARD_SOURCE_DIR}/init.h
     ${SOFACOMPONENTODESOLVERFORWARD_SOURCE_DIR}/EulerSolver.h
+    ${SOFACOMPONENTODESOLVERFORWARD_SOURCE_DIR}/EulerExplicitSolver.h
     ${SOFACOMPONENTODESOLVERFORWARD_SOURCE_DIR}/CentralDifferenceSolver.h
     ${SOFACOMPONENTODESOLVERFORWARD_SOURCE_DIR}/RungeKutta2Solver.h
     ${SOFACOMPONENTODESOLVERFORWARD_SOURCE_DIR}/RungeKutta4Solver.h
@@ -15,7 +16,7 @@ set(HEADER_FILES
 
 set(SOURCE_FILES
     ${SOFACOMPONENTODESOLVERFORWARD_SOURCE_DIR}/init.cpp
-    ${SOFACOMPONENTODESOLVERFORWARD_SOURCE_DIR}/EulerSolver.cpp
+    ${SOFACOMPONENTODESOLVERFORWARD_SOURCE_DIR}/EulerExplicitSolver.cpp
     ${SOFACOMPONENTODESOLVERFORWARD_SOURCE_DIR}/CentralDifferenceSolver.cpp
     ${SOFACOMPONENTODESOLVERFORWARD_SOURCE_DIR}/RungeKutta2Solver.cpp
     ${SOFACOMPONENTODESOLVERFORWARD_SOURCE_DIR}/RungeKutta4Solver.cpp
